refactored model and controller for create user
