+++ client/src/app/store.js
@@ -1,4 +1,5 @@
 import { configureStore } from "@reduxjs/toolkit";
+
 import authReducer from "../pages/Home/authSlice";
 
 export const store = configureStore({
