--- client/src/context.js
@@ -1,9 +1,0 @@
-import { useReducer } from "react";
-
-const reducer = (state, action) => {
-  return state;
-};
-
-const contextProvider = () => {
-  const [state, dispatch] = useReducer(reducer, {});
-};
