+++ client/src/pages/Login/Login.jsx
@@ -1,0 +1,108 @@
+import { Button, Checkbox, Form, Input } from "antd";
+import { useSelector, useDispatch } from "react-redux";
+import { useNavigate } from "react-router-dom";
+
+import LayoutOne from "/src/components/LayoutOne";
+import { login } from "../Home/authSlice";
+
+const Login = () => {
+  const loginState = useSelector((state) => state.auth.user);
+  const dispatch = useDispatch();
+  const navigate = useNavigate();
+
+  console.log("this is loginState", loginState);
+
+  const onFinish = async (credentials) => {
+    const { username, password } = credentials;
+    // console.log("submited:", credentials);
+    try {
+      const { payload } = await dispatch(login({ username, password }));
+      // console.log("dispatch payload", payload);
+
+      if (payload.success) {
+        navigate("/");
+      }
+    } catch (err) {
+      alert(err);
+    }
+  };
+
+  const onFinishFailed = (errorInfo) => {
+    console.log("Failed:", errorInfo);
+  };
+  return (
+    <LayoutOne>
+      <div className="login">
+        <h1>Login</h1>
+        <Form
+          name="basic"
+          labelCol={{
+            span: 8,
+          }}
+          wrapperCol={{
+            span: 16,
+          }}
+          style={{
+            maxWidth: 600,
+          }}
+          initialValues={{
+            remember: true,
+          }}
+          onFinish={onFinish}
+          onFinishFailed={onFinishFailed}
+          autoComplete="off"
+        >
+          <Form.Item
+            label="Username"
+            name="username"
+            rules={[
+              {
+                required: true,
+                message: "Please input your username!",
+              },
+            ]}
+          >
+            <Input />
+          </Form.Item>
+
+          <Form.Item
+            label="Password"
+            name="password"
+            rules={[
+              {
+                required: true,
+                message: "Please input your password!",
+              },
+            ]}
+          >
+            <Input.Password />
+          </Form.Item>
+
+          <Form.Item
+            name="remember"
+            valuePropName="checked"
+            wrapperCol={{
+              offset: 8,
+              span: 16,
+            }}
+          >
+            <Checkbox>Remember me</Checkbox>
+          </Form.Item>
+
+          <Form.Item
+            wrapperCol={{
+              offset: 8,
+              span: 16,
+            }}
+          >
+            <Button type="primary" htmlType="submit">
+              Submit
+            </Button>
+          </Form.Item>
+        </Form>
+      </div>
+    </LayoutOne>
+  );
+};
+
+export default Login;
